Set Up
Setting up the new frame to allow users to change the schedule used in
the simulation. Also modified Frame and ToolBar a little to accomidate
the new frame.

Currently having problems sending the selection to Frame (Or maybe it
should go to ToolBar?), will try to work on that later.